Add task buttons to the editor, remove range clamping of the typed value
Fix ranges not supporting floats
Start of making pipe custom presets to be a proper editor

+++ src/Objects/DataUtils.ts
@@ -48,7 +48,7 @@ export class DataUtils {
                     refValues.file = v.split(',')
                     break
                 case 'range':
-                    refValues.range = v.split(',').map(v => parseInt(v))
+                    refValues.range = v.split(',').map(v => parseFloat(v))
                     break
                 case 'code':
                     refValues.code = true
